Web UI: refactor into components + add search/filter, copy-to-clipboard, live SSE, conversation view
- Break monolithic 758-line App.tsx into 6 focused components:
  PlenaryList, PlenaryDetail, EventTimeline, PhaseStepper, StanceSummary, CopyButton
- Extract shared types (types.ts) and utilities (utils.ts)
- Add search bar and open/closed/all filter buttons to list view
- Add copy-to-clipboard buttons for plenary IDs and proposal IDs
- Add global SSE stream for live list updates (no more stale list)
- Add conversation view for event timeline (chat-style inline messages)
  with toggle to raw JSON view
- Fix event expansion to render inline (not in separate loop)
- Add keyboard accessibility (Enter/Space, ARIA) to event timeline
- Fix Vite proxy config for proper SSE streaming passthrough

Closes #27
